overhaul runtests structure -> SciML style (#36)
