+++ Project.toml
@@ -13,8 +13,8 @@ Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 [compat]
 CEnum = "0.4"
 OrderedCollections = "1.4"
-Unitful = "1.9"
 PrettyTables = "1.3"
+Unitful = "1.9"
 julia = "1"
 
 [extras]
@@ -24,7 +24,8 @@ Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "LinearAlgebra", "Printf", "CUDA", "Libdl", "Random", "IterTools"]
+test = ["Test", "LinearAlgebra", "Printf", "CUDA", "Libdl", "Random", "IterTools", "SafeTestsets"]
