+++ src/LIKWID.jl
@@ -81,10 +81,13 @@ function init(; gpu=false)
     init_numa()
     init_affinity()
     PerfMon.init()
-    HPM.init()
     Timer.init()
-    Power.init()
-    Freq.init()
+    if LIKWID.accessmode() == LIKWID.LibLikwid.ACCESSMODE_DAEMON
+        HPM.init()
+        Power.init()
+        # init_thermal(0)
+        Freq.init()
+    end
     if gpu && gpusupport()
         GPUMarker.init()
         init_topology_gpu()
